+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+#### Version 1.1.0 (2019-04-17)
+
+New Features:
+- Create topic automatically before sending/subscribing
+
+Bug Fixes:
+- Use shared client instead of creating one everytime, trying to fix `Failed to start Sarama producer: kafka: client has run out of available brokers to talk to (Is your cluster reachable?)`
+
 #### Version 1.0.0 (2019-04-15)
 
 New Features: